refactor: update extension.lua
