+++ install/asterisk/pjsip.conf
@@ -8,10 +8,9 @@ default_from_user = RBT-MEDIA
 [transport-udp]
 type = transport
 protocol = udp
-bind = 0.0.0.0
-
+bind = 0.0.0.0:5060
 
 [transport-tcp]
 type = transport
 protocol = tcp
-bind = 0.0.0.0
+bind = 0.0.0.0:5060
